Fixed bug #Issue 3
Completed bug fix and tests.
Description:
Multiple libraries where the user can search for the book. Only one has
got it.

If the user goes to the Semos library looking for the book there will be
no problems to complete the quest.
However, if they go to the library in Ados, the librarian won't know
anything about the book and the player will be confused.

Apparently, there are at least two libraries in the game. One is in
Semos and the other one is in Ados.
The library in Semos contains multiple books on the topics of mining and
related subjects.
There is a quest, the Obsidian Knife Quest, that asks you to obtain a
book (from a librarian) and give it to someone else (not relevant).
If the player is searching for the book in the Ados library, the
librarian won't know about the book and won't answer providing relevant
information.
There is no problem with the librarian at Semos.
The goal here is to make the librarian at Ados, whose name is
Wikipedian, provide an answer when asked about the book that the player
needs to obtain for the Quest.
The minimum, what the report asks for, is that Wikipedian redirects the
player to the Semos library.

The JUnit tests were failing due to a Null pointer. Referencing the npc
to Ceryl works fine, but this doesn't work for Wikipedian.
LibrarianNPC is Ceryl by default. Write the full path for the LibraryNPC
class in order to reference Wikipedian instead of Ceryl.