Fix bug in ActionsXMLLoader
There was a bug here caused by the parameters HashMap not being cleared
before processing the next action unless it also contained a parameters
element. This caused the parameters from a previous action to be copied
over to the next one if the next one did not have any parameters.